reporting/requirements.txt
psycopg2
pyyaml
streamlit
